requirements.txt
flet>=0.24.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pytest>=7.4.0
python-dotenv>=1.0.0
